CR 3266321: fix build with clang
During review or pull-request 62, we switched from using strlen() to
sizeof() for determining the lengths of the fixed strings used to
mark the begining/end of a PEM-encoded data type, since sizeof() is
evaluated at compile-time and strlen() might not be optimized out
by the compiler.  However, sizeof() returns a size_t, which is
unsigned, causing sign-comparison warnings from clang with
--strict-warnings (strlen() returns int).  Cast the sizeof() result
to silence the warnings.

Can be squashed into commit c406979c652b95a8d969388e0abb99f117936fe2 on rebase.
Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: to-be-published

+++ crypto/pem/pem_lib.c
@@ -951,9 +951,9 @@ static int sanitize_line(char *linebuf, int len, unsigned int flags)
 static char const beginstr[] = "-----BEGIN ";
 static char const endstr[] = "-----END ";
 static char const tailstr[] = "-----\n";
-#define BEGINLEN (sizeof(beginstr) - 1)
-#define ENDLEN (sizeof(endstr) - 1)
-#define TAILLEN (sizeof(tailstr) - 1)
+#define BEGINLEN ((int)sizeof(beginstr) - 1)
+#define ENDLEN ((int)sizeof(endstr) - 1)
+#define TAILLEN ((int)sizeof(tailstr) - 1)
 static int get_name(BIO *bp, char **name, unsigned int flags)
 {
     char *linebuf;
